+++ packages/block-library/src/editor.scss
@@ -15,6 +15,7 @@
 @import "./latest-comments/editor.scss";
 @import "./latest-posts/editor.scss";
 @import "./list/editor.scss";
+@import "./missing/editor.scss";
 @import "./more/editor.scss";
 @import "./nextpage/editor.scss";
 @import "./preformatted/editor.scss";
